pysemgrep: extract new scan_report.py from run_scan.py
run_scan.py was pretty big already, better have
small files with clear separation of concerns.

test plan:
make e2e
